+++ src/plone/app/fhirfield/indexes/es/mapping_types.py
@@ -143,7 +143,7 @@ HumanName = {
         'family': Token,
         'text': Text,
         'prefix': Token,
-        'period': Period,
+        'given': Token,
         'use': Token,
     },
 }